Move scatter to distributed. Add more tests
