+++ tt-train/sources/ttml/ttnn_fixed/distributed/ttnn_ops.hpp
@@ -1,0 +1,12 @@
+// SPDX-FileCopyrightText: (c) 2025 Tenstorrent AI ULC
+//
+// SPDX-License-Identifier: Apache-2.0
+
+#pragma once
+#include <core/ttnn_all_includes.hpp>
+
+namespace ttml::ttnn_fixed::distributed {
+
+tt::tt_metal::Tensor scatter(const tt::tt_metal::Tensor& t, int dim);
+
+}  // namespace ttml::ttnn_fixed::distributed
